Fix for HARMONY-743 ([classlib][security] compatibility: unexpected ArrayIndexOutOfBoundsException for MessageDigest.digest(byte[],Integer.MAX_VALUE,int);)
git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/classlib/trunk@420555 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/security/src/main/java/common/java/security/MessageDigest.java
@@ -165,7 +165,7 @@ public abstract class MessageDigest extends MessageDigestSpi {
      */
     public void update(byte[] input, int offset, int len) {
         if (input == null || offset < 0 || len < 0 ||
-                offset + len > input.length) {
+                (long)offset + (long)len > input.length) {
             throw new IllegalArgumentException(
                     "Incorrect offset/len parameters");
         }
@@ -201,7 +201,7 @@ public abstract class MessageDigest extends MessageDigestSpi {
      */
     public int digest(byte[] buf, int offset, int len) throws DigestException {
         if (buf == null || offset < 0 || len < 0 ||
-                offset + len > buf.length) {
+                (long)offset + (long)len > buf.length) {
             throw new IllegalArgumentException(
                     "Incorrect offset/len parameters");
         }
